rough cuts
bulk move over the socorro common folders, build process, tests, etc

now we find and cut and all the strange knots looped around other parts of the application -- unittests that have mutated into integration tests because they needed an error class from a library that talks to the database when its imported, or static methods stubbing a generic api that were hacked up with hardcoded values to make the api correct but defer implementation until later.

The goal here is to find the smallest subset of stuff that can be implemented as a stand alone library. At a first pass, this looks like:

  - socorro.app
  - socorro.lib
  - socorro.unittest.app
  - socorro.unittest.lib

The external classes may have to get pulled in, but with them comes some pretty heavy integration testing and for sanity's sake we'd be better off avoiding it for now.

Some conflicts we can fake, some we will have to alter the primary socorro repo to work around when we finally go to import this as a lib.