simply break overly strong ties to socorro classes
The chicken and egg problem presented by the application defaults was solved in a way that causes the whole class to load in testing. This patch breaks that in perhaps an overly naive way, by removing any default associations.

It may be better to push these defaults to the apps themselves, and make each responsible for their own defaults.