temporarily shim default class associations
Abbreviated names for full qualified class paths are used all over the socorro app and automation. This restores the original shim, without the tests to avoid a circular dep against the socorro code.

This should still be removed and replaced with an actual service discovery mechanism or obviated entirely some other way. Classes cannot define their own shortcuts since the class has to be loaded to express that variable. Might work if we crawl the all classes or something similar.

Alternatively, we could replace all shortcut paths with fully qualified paths. This means <100 references across the repository and automation, maybe closer to 50.