deprecated
socorrolib was started with the intent of being a common library that would enable splitting the monorepo into microservices. Six months into that effort we've switched approaches and are rewriting new services rather than trying to split out the extant modules.

with that decision made, having this independent of the main repository makes maintenance unecessarily harder. it has been folded back into the monorepo and this library is now deprecated.